_archive/requirements.txt
fal-client>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
